Chapter 7.
Chapter 7 deals with interfaces, which to my most basic understanding of
this matter seem to be a way of defining types of types. For example,
the type constraint `Eq ty` might be viewed as a type judgement,
claiming the type `type` to be of type `Eq`.

In any case, interfaces allow you to introduce hierarchies between
types, whereas formerly we only considered types as isolated
individuals.