Chapter 8 completed.
Many people don't like considering "trivial" edge cases, like empty
sets, 0x0 matrices, or other things you "never need in real life". Too
bad! Not only because it often illuminates the general concept and
sometimes tells you that your definition is insufficient, more
importantly because it's generally a lot of fun!

The empty type---the main protagonist of section 8.3---is no exception
to this rule. In particular, we learn how a data type that has
absolutely no constructors for it can be used to revamp the type of
`checkEqNat` in such a way that it completely fixes the semantics.

Yeah, empty types rock!