Chapter 10: Views.
This chapter introduces views and the `with` construction, which
together make it possible to pattern match on more complicated
expressions than just (iterated) constructors.

In section 10.1 we consider two examples, a view `ListLast` that allows
one to easily traverse a list from back-to-front, and a view `SplitList`
that allows one to pattern match on lists that are given as composite
`left ++ right` of two sub-lists. The latter view is useful for
implementing the "merge sort" algorithm.