Finished chapter 10.
In this chapter we learned that "views" are an amazingly useful concept
that allows you to

- formalize various ways in which you can take apart
a data structure (e.g. traversing a list from back to front, viewing
a list of size n*m as a list of lists etc.)
- pattern match on expressions more complicated than iterated
constructors in function definitions
- expose a stable interface to a data type whose constructors are
private (and subject to change)
- ensure totality of functions that use pattern matching on views by
using *recursive views*

TL;DR: Views are awesome!