Chapter 12: State.
Just like you can use monads to have i/o in a pure language by
separating the *description* of a sequences of actions from their
*execution*, you can use monads---in this case the `State` monad---to
separate the description of a sequence of state mutations from their
execution.

In the first section of chapter 12, we see how useful this can be by
implementing a function that labels a tree. Without using the `State`
monad, you need to explicitly pass the state around which makes the code
look verbose and ugly. Using the `State` monad, it look about as slick
as it would if coded in C.