Chapter 13: State machines.
In the first section of chapter 13, we see how having dependend types in
the "state monad pattern" makes it possible to have state transitions be
restricted to according to some abstract state machine model.

We see two concrete examples, an elementary model of a door with a door
bell and a more delicate model of a vending machine.